Ignore dollar properties on merge
This improvement removes dollar properties from the change detection
logic. A reasonable amount of work has happened to help this be
implemented, mostly cleaning the repositories types, updating
dependencies, etc, but the primary fix is relatively simple.

Prior to calculating the diff we now recursively strip the properties
beginning with "$" from the objects.